When code saving is enabled, put path to program file in config so that the frontend can find it.
